LCAM-842: Added data contract for generic calculate contribution. Refactored the existing maat specific logic
