+++ crown-court-contribution/src/main/java/uk/gov/justice/laa/crime/contribution/builder/CalculateContributionResponseBuilder.java
@@ -3,7 +3,7 @@ package uk.gov.justice.laa.crime.contribution.builder;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.springframework.stereotype.Component;
-import uk.gov.justice.laa.crime.contribution.model.CalculateContributionResponse;
+import uk.gov.justice.laa.crime.contribution.model.MaatCalculateContributionResponse;
 import uk.gov.justice.laa.crime.contribution.model.Contribution;
 
 import static uk.gov.justice.laa.crime.contribution.util.DateUtil.convertDateToDateTime;
@@ -13,35 +13,35 @@ import static uk.gov.justice.laa.crime.contribution.util.DateUtil.getLocalDateSt
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class CalculateContributionResponseBuilder {
 
-    public static CalculateContributionResponse build(Contribution contribution) {
-        CalculateContributionResponse calculateContributionResponse = new CalculateContributionResponse();
-        calculateContributionResponse.withId(contribution.getId());
-        calculateContributionResponse.withApplId(contribution.getApplId());
-        calculateContributionResponse.withRepId(contribution.getRepId());
-        calculateContributionResponse.withContributionFileId(contribution.getContributionFileId());
-        calculateContributionResponse.withEffectiveDate(getLocalDateString(contribution.getEffectiveDate()));
-        calculateContributionResponse.withCalcDate(convertDateToDateTime(contribution.getCalcDate()));
-        calculateContributionResponse.withContributionCap(contribution.getContributionCap());
-        calculateContributionResponse.withMonthlyContributions(contribution.getMonthlyContributions());
-        calculateContributionResponse.withUpfrontContributions(contribution.getUpfrontContributions());
-        calculateContributionResponse.withUpliftApplied(contribution.getUpliftApplied());
-        calculateContributionResponse.withBasedOn(contribution.getBasedOn());
-        calculateContributionResponse.withTransferStatus(contribution.getTransferStatus());
-        calculateContributionResponse.withDateUpliftApplied(convertDateToDateTime(contribution.getDateUpliftApplied()));
-        calculateContributionResponse.withDateUpliftRemoved(convertDateToDateTime(contribution.getDateUpliftRemoved()));
-        calculateContributionResponse.withDateCreated(contribution.getDateCreated());
-        calculateContributionResponse.withUserCreated(contribution.getUserCreated());
-        calculateContributionResponse.withDateModified(contribution.getDateModified());
-        calculateContributionResponse.withUserModified(contribution.getUserModified());
-        calculateContributionResponse.withCreateContributionOrder(contribution.getCreateContributionOrder());
-        calculateContributionResponse.withCorrespondenceId(contribution.getCorrespondenceId());
-        calculateContributionResponse.withActive(contribution.getActive());
-        calculateContributionResponse.withReplacedDate(convertDateToDateTime(contribution.getReplacedDate()));
-        calculateContributionResponse.withLatest(contribution.getLatest());
-        calculateContributionResponse.withCcOutcomeCount(contribution.getCcOutcomeCount());
-        calculateContributionResponse.withSeHistoryId(contribution.getSeHistoryId());
+    public static MaatCalculateContributionResponse build(Contribution contribution) {
+        MaatCalculateContributionResponse maatCalculateContributionResponse = new MaatCalculateContributionResponse();
+        maatCalculateContributionResponse.withId(contribution.getId());
+        maatCalculateContributionResponse.withApplId(contribution.getApplId());
+        maatCalculateContributionResponse.withRepId(contribution.getRepId());
+        maatCalculateContributionResponse.withContributionFileId(contribution.getContributionFileId());
+        maatCalculateContributionResponse.withEffectiveDate(getLocalDateString(contribution.getEffectiveDate()));
+        maatCalculateContributionResponse.withCalcDate(convertDateToDateTime(contribution.getCalcDate()));
+        maatCalculateContributionResponse.withContributionCap(contribution.getContributionCap());
+        maatCalculateContributionResponse.withMonthlyContributions(contribution.getMonthlyContributions());
+        maatCalculateContributionResponse.withUpfrontContributions(contribution.getUpfrontContributions());
+        maatCalculateContributionResponse.withUpliftApplied(contribution.getUpliftApplied());
+        maatCalculateContributionResponse.withBasedOn(contribution.getBasedOn());
+        maatCalculateContributionResponse.withTransferStatus(contribution.getTransferStatus());
+        maatCalculateContributionResponse.withDateUpliftApplied(convertDateToDateTime(contribution.getDateUpliftApplied()));
+        maatCalculateContributionResponse.withDateUpliftRemoved(convertDateToDateTime(contribution.getDateUpliftRemoved()));
+        maatCalculateContributionResponse.withDateCreated(contribution.getDateCreated());
+        maatCalculateContributionResponse.withUserCreated(contribution.getUserCreated());
+        maatCalculateContributionResponse.withDateModified(contribution.getDateModified());
+        maatCalculateContributionResponse.withUserModified(contribution.getUserModified());
+        maatCalculateContributionResponse.withCreateContributionOrder(contribution.getCreateContributionOrder());
+        maatCalculateContributionResponse.withCorrespondenceId(contribution.getCorrespondenceId());
+        maatCalculateContributionResponse.withActive(contribution.getActive());
+        maatCalculateContributionResponse.withReplacedDate(convertDateToDateTime(contribution.getReplacedDate()));
+        maatCalculateContributionResponse.withLatest(contribution.getLatest());
+        maatCalculateContributionResponse.withCcOutcomeCount(contribution.getCcOutcomeCount());
+        maatCalculateContributionResponse.withSeHistoryId(contribution.getSeHistoryId());
 
-        return calculateContributionResponse;
+        return maatCalculateContributionResponse;
     }
 
 }
