+++ crown-court-contribution/src/main/java/uk/gov/justice/laa/crime/contribution/builder/ContributionDTOBuilder.java
@@ -4,7 +4,7 @@ import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.springframework.stereotype.Component;
 import uk.gov.justice.laa.crime.contribution.dto.CalculateContributionDTO;
-import uk.gov.justice.laa.crime.contribution.model.CalculateContributionRequest;
+import uk.gov.justice.laa.crime.contribution.model.MaatCalculateContributionRequest;
 import uk.gov.justice.laa.crime.contribution.model.Contribution;
 import uk.gov.justice.laa.crime.contribution.model.CreateContributionRequest;
 import uk.gov.justice.laa.crime.contribution.util.DateUtil;
@@ -13,7 +13,7 @@ import uk.gov.justice.laa.crime.contribution.util.DateUtil;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class ContributionDTOBuilder {
 
-    public static CalculateContributionDTO build(final CalculateContributionRequest request) {
+    public static CalculateContributionDTO build(final MaatCalculateContributionRequest request) {
         CalculateContributionDTO.CalculateContributionDTOBuilder builder = CalculateContributionDTO.builder()
                 .id(request.getRepId())
                 .applId(request.getApplId())
