+++ crown-court-contribution/src/main/resources/schemas/apiCalculateContributionResponse.json
@@ -1,0 +1,30 @@
+{
+  "$schema": "http://json-schema.org/draft-04/schema#",
+  "id": "apiCalculateContributionResponse.json",
+  "type": "object",
+  "title": "Calculate Contribution Response",
+  "description": "Data Contract for Calc Contribs Response",
+  "properties": {
+    "monthlyContributions": {
+      "type": "number",
+      "default": 0.0,
+      "description": "The monthly contribution amount"
+    },
+    "upfrontContributions": {
+      "type": "number",
+      "default": 0.0,
+      "description": "The upfront contribution amount"
+    },
+    "upliftApplied": {
+      "type": "string",
+      "default": "N",
+      "description": "Indicates if an income uplift was applied"
+    },
+    "basedOn": {
+      "type": ["string",null],
+      "default": null,
+      "description": "Indicates what the contributions are based on (e.g. Means assessment)"
+    }
+  },
+  "additionalProperties": false
+}
